refactor: explicit imports and MUI entry usage outside stories
Made-with: Cursor
