+++ packages/open-ui-kit/src/components/filters/components/filters-bar/filters-bar-actions.tsx
@@ -14,7 +14,7 @@ import { Box, Button, Stack } from "@mui/material";
 import { useEffect, useState } from "react";
 import { styles } from "./styles";
 import { greyPalette } from "@/theme/color-palette";
-import { SearchField } from "@/components";
+import { SearchField } from "@/components/search-field";
 
 export interface FiltersBarActionsProps {
   onFiltersButtonClick?: () => void;
